move to a single EmptyState component with variants
