--- packages/grafana-ui/src/components/EmptyState/EmptySearchState/EmptySearchState.mdx
@@ -1,37 +1,0 @@
-import { ArgTypes } from '@storybook/blocks';
-import { EmptySearchState } from './EmptySearchState';
-
-# EmptySearchState
-
-## When to use
-
-Use in place of a results table or list when the search query returns no results.
-
-The component displays a message and an image to provide visual confirmation to a user that the search has completed correctly.
-
-## Usage
-
-### Basic
-
-There are sensible defaults for the image and message, so in most cases you can use it without any additional props.
-
-```jsx
-import { EmptySearchState } from '@grafana/ui';
-
-<EmptySearchState />;
-```
-
-### Providing a custom message, image, or additional information
-
-```jsx
-import { Button, EmptySearchState } from '@grafana/ui';
-
-<EmptySearchState image={<img src="<pathToImage>" />} message="No playlists found">
-  // Optionally provide some additional information here. You could even provide a button to clear the search query.
-  <Button variant="secondary" onClick={clearSearchQuery} />
-</EmptySearchState>;
-```
-
-## Props
-
-<ArgTypes of={EmptySearchState} />
