--- packages/grafana-ui/src/components/EmptyState/EmptySearchState/EmptySearchState.story.tsx
@@ -1,33 +1,0 @@
-import { Meta, StoryFn } from '@storybook/react';
-import React from 'react';
-
-import { EmptySearchState } from './EmptySearchState';
-import mdx from './EmptySearchState.mdx';
-
-const meta: Meta<typeof EmptySearchState> = {
-  title: 'General/EmptyState/EmptySearchState',
-  component: EmptySearchState,
-  parameters: {
-    docs: {
-      page: mdx,
-    },
-    controls: {
-      exclude: ['image'],
-    },
-  },
-  argTypes: {
-    children: {
-      type: 'string',
-    },
-  },
-};
-
-export const Basic: StoryFn<typeof EmptySearchState> = (args) => {
-  return <EmptySearchState {...args} />;
-};
-
-Basic.args = {
-  children: 'Use this space to add any additional information',
-};
-
-export default meta;
