+++ public/locales/pseudo-LOCALE/grafana.json
@@ -563,8 +563,8 @@
     "drawer": {
       "close": "Cľőşę"
     },
-    "empty-search-state": {
-      "message": "Ńő řęşūľŧş ƒőūŉđ"
+    "empty-state": {
+      "search-message": "Ńő řęşūľŧş ƒőūŉđ"
     },
     "modal": {
       "close-tooltip": "Cľőşę"
